requirements.txt
aiogram==3.4.1
openai>=1.30.0
python-dotenv>=1.0.1
aiohttp>=3.8.0
aiofiles>=23.0.0
